Something worked
This is bad form probably, but I'm trying my hardest to get the protobuf gradle plugin to generate kotlin sources and I got it to at least generate the java sources. As far as I understand, I need these for kotlin sources to be generated and I don't want to lose this progress.